Removed the qr screens and stateselector from the diagnosis folder and moved them into qr and news folders respectively in the screen folder. Also renamed the diagnosis navigator to be the news navigator.
