+++ screens/qr/QRResultsScreen.js
