+++ screens/qr/QRScanningScreen.js
